Break Linux copy function out to own mod for sparse refactoring.
